+++ packages/cli/src/lint/lintSchema.ts
@@ -407,14 +407,9 @@ export const lintSchema = async (noLog = false) => {
                                 }
                             }
                             if (relation.loadAll) {
-                                if (relationCollection.parentCollection) {
+                                if (!["OneToOne", "OneToMany"].includes(relationField.type)) {
                                     errors.push(
-                                        `Collection ${collectionName} has a relation list for collection ${relation.collection} that has loadAll enabled but is on a subcollection`,
-                                    )
-                                }
-                                if (relationCollection.indexExemption) {
-                                    errors.push(
-                                        `Collection ${collectionName} has a relation list for collection ${relation.collection} that has loadAll enabled but is on a collection with indexExemption enabled`,
+                                        `Collection ${collectionName} has a relation list for collection ${relation.collection} that has loadAll enabled but is not a one to one or one to many field`,
                                     )
                                 }
                             }